bee_backend/win32: buffer child ouput without temp files
More and more often we have trouble with windows defender interfering
with tundra's temp files, which it uses to buffer child process output.
Maybe if we use pipes instead of files, we won't have this
trouble.

Using async/overlapped IO will come in handy if/when we refactor to not
use an os thread per build slot (in order to scale up to -j 1000).

https://jira.unity3d.com/browse/DS-2646